--- src/main/java/com/lolzorrior/supernaturalmod/capabilities/PowerFactory.java
@@ -1,11 +1,0 @@
-package com.lolzorrior.supernaturalmod.capabilities;
-
-
-import java.util.concurrent.Callable;
-
-public class PowerFactory implements Callable<ISupernaturalPower> {
-    @Override
-    public ISupernaturalPower call() throws Exception {
-        return new SupernaturalPower();
-    }
-}
