+++ src/main/java/com/lolzorrior/supernaturalmod/capabilities/SupernaturalPowerFactory.java
@@ -1,0 +1,11 @@
+package com.lolzorrior.supernaturalmod.capabilities;
+
+
+import java.util.concurrent.Callable;
+
+public class SupernaturalPowerFactory implements Callable<ISupernaturalPower> {
+    @Override
+    public ISupernaturalPower call() throws Exception {
+        return new SupernaturalPower();
+    }
+}
